==> 01_first attempt/01_02_Added_requests.ipynb <==

import pandas as pd
import numpy as np

Taking requests that we havent taken in yet

with requests as (
select distinct requestid as requestid from airtable_contracts
where grosscontractsigned = 1
or grosscontractsignedfreelancer = 1
),
ce as ( select distinct requestid as id from calendar_event where createdat > '2025-03-01'
 and createdat < '2025-07-15'
and type <> 'VACATION'
)
select r.requestid from requests r left outer join ce on ce.id = r.requestid
WHERE ce.id IS NULL

view --> new_requests

adding leadid

select n.requestid, r.leadid from new_requests n join request r on r.id = n.requestid )

view --> info1


-- download of 9/4

path = r"C:\Users\ElifYilmaz\Downloads\recover contracts.csv"
df = pd.read_csv(path)
df.columns = df.columns.str.lower()
