==> 01_first attempt/03_02_leakage handling.ipynb <==
import pandas as pd

lets try taking data that ois only with net sc1
from 19k we drop to 12k

df = pd.read_csv('processed_data/merged_df.csv')
df.columns=df.columns.str.lower()
df.info()
df["grosscontractsigned"].value_counts()

df = (df[df['net_sc'] != 0]).copy()

df.info()
df["grosscontractsigned"].value_counts()

# only made it to net sc 
# Save your merged dataframe
df.to_csv('processed_data/merged_df_net_sc.csv', index=False)

print("Dataset saved successfully!")
